# contract/openapi.yaml
openapi: 3.1.0
info:
  title: AI 健身助手 P0 API
  version: 1.0.0
  description: >-
    微信首发但渠道无关的业务契约。关键训练数字由版本化确定性规则产生，
    AI 仅提供候选、排序和解释；用户归属始终由认证主体确定。
servers:
  - url: /
    description: local/test relative base
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/workout-sessions:
    post:
      tags: [Workouts]
      operationId: startWorkoutSession
      parameters:
        - {$ref: './schemas/common.yaml#/components/parameters/IdempotencyKey'}
      requestBody: {$ref: '#/components/requestBodies/WorkoutSessionStart'}
      responses:
        '201':
          description: 会话已创建；同一用户和 clientSessionKey 仅在原会话仍为 CREATED 时返回首次创建结果
          content:
            application/json:
              schema: {$ref: './schemas/workout.yaml#/components/schemas/WorkoutSessionResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '409':
          description: 服务端原子重算后需要恢复确认、幂等键冲突、已有未结束训练，或该启动键对应会话已经结束；均不会创建第二个会话。
          content:
            application/json:
              schema: {$ref: './schemas/workout.yaml#/components/schemas/WorkoutSessionStartConflictResponse'}
        default: {$ref: '#/components/responses/DefaultError'}
    get:
      tags: [Workouts]
      operationId: listWorkoutSessions
      parameters:
        - {$ref: './schemas/common.yaml#/components/parameters/Cursor'}
        - {name: limit, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 50, default: 20}}
      responses:
        '200':
          description: 当前用户的终态训练历史，使用稳定不透明游标分页
          content:
            application/json:
              schema: {$ref: './schemas/workout.yaml#/components/schemas/WorkoutHistoryResponse'}
        default: {$ref: '#/components/responses/DefaultError'}
  /api/v1/workout-sessions/{id}:
    get:
      tags: [Workouts]
      operationId: getWorkoutSession
      parameters:
        - {$ref: '#/components/parameters/IdPath'}
      responses:
        '200':
          description: 用户归属的不可变训练快照
          content:
            application/json:
              schema: {$ref: './schemas/workout.yaml#/components/schemas/WorkoutSessionResponse'}
        '404': {$ref: '#/components/responses/NotFound'}
        default: {$ref: '#/components/responses/DefaultError'}
  /api/v1/workout-sessions/{id}/status:
    put:
      tags: [Workouts]
      operationId: updateWorkoutSessionStatus
      parameters:
        - {$ref: '#/components/parameters/IdPath'}
      requestBody: {$ref: '#/components/requestBodies/WorkoutStatusUpdate'}
      responses:
        '200':
          description: 状态转换后的会话
          content:
            application/json:
              schema: {$ref: './schemas/workout.yaml#/components/schemas/WorkoutSessionResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '409': {$ref: '#/components/responses/VersionConflict'}
        default: {$ref: '#/components/responses/DefaultError'}
  /api/v1/workout-sessions/{id}/exercises/{exerciseId}:
    put:
      tags: [Workouts]
      operationId: updateWorkoutSessionExercise
      parameters:
        - {$ref: '#/components/parameters/IdPath'}
        - {$ref: '#/components/parameters/ExerciseIdPath'}
      requestBody: {$ref: '#/components/requestBodies/WorkoutExerciseUpdate'}
      responses:
        '200':
          description: 仅当前训练生效的替换覆盖层；原计划版本和原始训练快照保持不变
          content:
            application/json:
              schema: {$ref: './schemas/workout.yaml#/components/schemas/WorkoutSessionResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '409':
          description: 版本冲突，或当前训练已无不重复且安全的替代动作
          content:
            application/json:
              schema: {$ref: './schemas/workout.yaml#/components/schemas/WorkoutExerciseUpdateConflictResponse'}
        default: {$ref: '#/components/responses/DefaultError'}
...
  /api/v1/workout-sessions/{id}/complete:
    post:
      tags: [Workouts]
      operationId: completeWorkoutSession
      parameters:
        - {$ref: '#/components/parameters/IdPath'}
        - {$ref: './schemas/common.yaml#/components/parameters/IdempotencyKey'}
      requestBody: {$ref: '#/components/requestBodies/WorkoutCompletion'}
      responses:
        '200':
          description: 幂等训练结算；提前结束保留已完成组且不自动进入加重
          content:
            application/json:
              schema: {$ref: './schemas/workout.yaml#/components/schemas/WorkoutCompletionResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '409': {$ref: '#/components/responses/VersionConflict'}
        default: {$ref: '#/components/responses/DefaultError'}
...
  /api/v1/workout-sessions/{id}/summary:
    get:
      tags: [Workouts]
      operationId: getWorkoutSessionSummary
      parameters:
        - {$ref: '#/components/parameters/IdPath'}
      responses:
        '200':
          description: 已完成或提前结束训练的汇总事实
          content:
            application/json:
              schema: {$ref: './schemas/workout.yaml#/components/schemas/WorkoutSessionSummaryResponse'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409':
          description: 训练尚未进入终态，错误码为 WORKOUT_NOT_TERMINAL。
          content:
            application/json:
              schema: {$ref: './schemas/workout.yaml#/components/schemas/WorkoutNotTerminalErrorResponse'}
        default: {$ref: '#/components/responses/DefaultError'}
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    IdPath: {name: id, in: path, required: true, schema: {type: string}}
...
    ExerciseIdPath: {name: exerciseId, in: path, required: true, schema: {type: string}}
...
  requestBodies:
...
    WorkoutSessionStart: {required: true, content: {application/json: {schema: {$ref: './schemas/workout.yaml#/components/schemas/StartWorkoutSessionRequest'}}}}
    WorkoutStatusUpdate: {required: true, content: {application/json: {schema: {$ref: './schemas/workout.yaml#/components/schemas/UpdateSessionStatusRequest'}}}}
    WorkoutExerciseUpdate: {required: true, content: {application/json: {schema: {$ref: './schemas/workout.yaml#/components/schemas/UpdateSessionExerciseRequest'}}}}
...
    WorkoutCompletion: {required: true, content: {application/json: {schema: {$ref: './schemas/workout.yaml#/components/schemas/CompleteSessionRequest'}}}}
...
  responses:
...
    BadRequest: {$ref: './schemas/common.yaml#/components/responses/BadRequest'}
    Unauthorized: {$ref: './schemas/common.yaml#/components/responses/Unauthorized'}
    NotFound: {$ref: './schemas/common.yaml#/components/responses/NotFound'}
    VersionConflict: {$ref: './schemas/common.yaml#/components/responses/VersionConflict'}
...
    DefaultError: {$ref: './schemas/common.yaml#/components/responses/DefaultError'}